requirements.txt
pandas
numpy
scikit-learn
tensorflow
matplotlib
pyyaml
